not fixed but yes

+++ src/main/java/got/common/enchant/GOTEnchantmentProtectionFall.java
@@ -10,7 +10,7 @@ public class GOTEnchantmentProtectionFall extends GOTEnchantmentProtectionSpecia
 
 	@Override
 	public int calcIntProtection() {
-		float f = (float) protectLevel * (float) (protectLevel + 1) / 2.0f;
+		float f = protectLevel * (protectLevel + 1) / 2.0F;
 		return 3 + MathHelper.floor_float(f);
 	}
